% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSq.R
\name{ChiSq.Bins}
\alias{ChiSq.Bins}
\alias{ChiSq.pois}
\title{Chi Square}
\usage{
ChiSq.Bins(
  bin.bounds,
  ObsFreq,
  mean,
  sd,
  NullDist = c("normal", "uniform"),
  df = length(ObsFreq) - 1
)

ChiSq.pois(
  counts,
  ObsFreq,
  lambda = sum(counts * ObsFreq)/sum(ObsFreq),
  boundary = c("neither", "greater", "less"),
  df = length(ObsFreq) - 1
)
}
\arguments{
\item{bin.bounds}{A numeric vector for the boundaries of the bins. -Inf and positive Inf will automatically be added to each end.}

\item{mean}{The sample x-value  mean of the data}

\item{sd}{The standard deviation}

\item{NullDist}{The expected distribution under the null hypothesis. Currently supports normal and uniform}

\item{df}{Degrees of freedom}

\item{counts}{A numeric vector of the frequencies of occurence within the time unit.}

\item{lambda}{The parameter of the Poission distribution; will be calculated if not provided.}

\item{Obs.Freq}{A numeric vector of the time units for each count.}
}
\description{
This function produces a table and runs a Chi Square test on counts of binned data or Poisson Ditrsibutions.
}
\section{Functions}{
\itemize{
\item \code{ChiSq.Bins}: Chi-Square for Binned Data

\item \code{ChiSq.pois}: Chi-Square for Binned Data
}}

