% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{constants}
\alias{constants}
\title{Constants}
\usage{
constants(x)
}
\arguments{
\item{x}{name of constant}
}
\description{
This function loads a constant.
}
\section{Constants}{

\itemize{
\item{"h"}{ --- Planck constant (J\*m)}
\item{"c"}{ --- speed of light in a vaccum (m/s)}
\item{"e"}{ --- Euler number}
\item{"R"}{ --- ideal gas constant (J/(k\\emph{m))}
\item{"me"}{ --- mass of an electron (kg)}
\item{"F"}{ --- Faraday constant (C/mol)}
\item{"An"}{ --- Avogadro number}
\item{"Rh"}{ --- Rydbherg for hjydrogen (1/m)}
\item{"Cw"}{ --- molar heat capacity of water (J/(mol}K))}
}
}

