% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_as_tibble.R
\name{csv_as_tibble}
\alias{csv_as_tibble}
\title{CSV to Tibble}
\usage{
csv_as_tibble(x, n)
}
\arguments{
\item{x}{.csv file}

\item{n}{desired object variable}
}
\description{
This function takes a .csv file and assigns it as a tibble.
}
