% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_n.R
\name{list_n}
\alias{list_n}
\title{List with Names}
\usage{
list_n(x = 1:100)
}
\description{
This function converts a vector into a list with names eqaul to the values.
Useful for lapply and sapply operations.
}
