% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between.R
\name{pbetween.norm}
\alias{pbetween.norm}
\alias{pbetween.pois}
\alias{pbetween.t}
\alias{pbetween.binom}
\title{Probability Between}
\usage{
pbetween.norm(q1, q2, mean = 0, sd = 1)

pbetween.pois(q1, q2, l)

pbetween.t(q1, q2, df)

pbetween.binom(q1, q2, size, prob)
}
\description{
This function calculates probability between two qauntiles.
}
\section{Functions}{
\itemize{
\item \code{pbetween.norm}: Normal

\item \code{pbetween.pois}: Poisson

\item \code{pbetween.t}: Student's t

\item \code{pbetween.binom}: Binomial
}}

