% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pH approximation fucntion.R
\name{pH.approx}
\alias{pH.approx}
\title{pH approximator}
\usage{
pH.approx(Ka.base, Ka.exp, conc)
}
\description{
This function approximates the pH of a solution whre loss of base is negligible.
}
