% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolvar.R
\name{pool.var}
\alias{pool.var}
\alias{pool.sd}
\title{Pooled Standard Deviation and Variance}
\usage{
pool.var(n1, s1, n2, s2)

pool.sd(n1, s1, n2, s2)
}
\description{
Pooled Standard Deviation and Variance
}
\section{Functions}{
\itemize{
\item \code{pool.var}: Pooled Variance

\item \code{pool.sd}: Pooled Standard Deviation
}}

