% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power_func.z}
\alias{power_func.z}
\alias{power_func.t}
\title{Power Functions}
\usage{
power_func.z(
  mus,
  mu.0,
  sd,
  n,
  conf.level = 0.95,
  alternative = c("two.sided", "less", "greater")
)

power_func.t(
  mus,
  mu.0,
  sd,
  n,
  conf.level = 0.95,
  samples = c(1, 2),
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{mus}{a list or vector of true means or differences to be tested}

\item{mu.0}{the hypothesized mean or difference}

\item{sd}{standard deviation}

\item{n}{number of observations}

\item{alternative}{"two.sided", "less", or "greater", abbreviations accepted}

\item{samples}{number of samples (1 or 2)}
}
\description{
These functions create a tibble showing the power at different true means.
}
\section{Functions}{
\itemize{
\item \code{power_func.z}: creates a power function table based on the z method

\item \code{power_func.t}: creates a power function table based on the t method
}}

