% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Score}
\usage{
score(x, mu, sd, n)
}
\arguments{
\item{x}{value}

\item{mu}{mean}

\item{sd}{standard deviation}

\item{n}{number of obervations}
}
\description{
Calculates t-score or z-score
}
