% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_from_summ.R
\name{t_from_summ}
\alias{t_from_summ}
\title{t-Test from Summary Statistics}
\usage{
t_from_summ(
  m1,
  m2,
  s1,
  s2,
  n1,
  n2,
  alternative = c("two.sided", "less", "greater"),
  mu = 0,
  paired = FALSE,
  var.equal = FALSE,
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{m1}{mean of x}

\item{m2}{mean of y}

\item{s1}{standard deviation of x}

\item{s2}{standard deviation of y}

\item{n1}{number of observations in x}

\item{n2}{number of observations in y}
}
\description{
This is the t.test fucntion from the stats package adapted to work from summary statistics instead of vectors.
}
