% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_of_r.R
\name{t_of_r}
\alias{t_of_r}
\title{T-score of a Pearson correlation coefficient}
\usage{
t_of_r(r, n)
}
\arguments{
\item{r}{Pearson correlation coefficient}

\item{n}{number of observations}
}
\description{
T-score of a Pearson correlation coefficient
}
