% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd_nav.R
\name{wd_up}
\alias{wd_up}
\alias{wd_down}
\alias{wd_nav}
\title{Working Directory Navigators}
\usage{
wd_up(n_lev = 1)

wd_down(x)

wd_nav(up = 0, down)
}
\arguments{
\item{n_lev}{levels to navigate up, default 1}

\item{x}{character string for the downward file path}

\item{up}{levels to navigate up, default 0}

\item{down}{character string for the downward file path}
}
\description{
These functions allow the working directory to be moved up or down in the file path.
}
\section{Functions}{
\itemize{
\item \code{wd_up}: navigate wd up

\item \code{wd_down}: navigates wd down

\item \code{wd_nav}: navigates wd up, then down
}}

