% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between.R
\name{pbetween}
\alias{pbetween}
\title{Probability between}
\usage{
pbetween(q1, q2, m = 0, s = 1, l, size, p, df, method = "nomral")
}
\description{
This fucntion calculates probability between two qauntiles.
Method can be set to "norm" for normal, "pois" for Poisson, "binom" for binomial, or "t" for t.
}
